Decouple mol comparison and resonance from mol
