+++ rdmc/resonance/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+from rdmc.resonance.resonance import generate_radical_resonance_structures
